requirements.txt
asgiref==3.7.2
Authlib==1.2.1
Brotli==1.0.9
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
coverage==7.2.7
cryptography==41.0.2
dj-database-url==2.0.0
Django==4.2.2
django-bmemcached==0.3.0
django-cors-headers==4.1.0
django-environ==0.10.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
drf-jwt==1.19.2
ecdsa==0.18.0
exceptiongroup==1.1.1
Faker==19.2.0
gunicorn==20.1.0
idna==3.4
iniconfig==2.0.0
Markdown==3.4.3
packaging==23.1
pluggy==1.2.0
psycopg2-binary==2.9.6
pyasn1==0.5.0
pycparser==2.21
PyJWT==1.7.1
pytest==7.4.0
pytest-cov==4.1.0
pytest-django==4.5.2
python-binary-memcached==0.31.2
python-dateutil==2.8.2
python-decouple==3.8
python-jose==3.3.0
python-memcached==1.59
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
responses==0.23.1
rsa==4.9
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
types-PyYAML==6.0.12.11
typing_extensions==4.7.0
uhashring==2.3
urllib3==2.0.3
whitenoise==6.5.0
